+++ src/frontend/apps/impress/src/features/docs/doc-header/components/index.ts
@@ -1,3 +1,3 @@
 export * from './DocHeader';
 export * from './DocTitle';
-export * from './FloatingBar';
+export * from './DocFloatingBar';
